# backend/docs/swagger.yaml
basePath: /
definitions:
  res.ErrorConstant:
    properties:
      code:
        type: integer
      data: {}
      error_fields:
        items:
          $ref: '#/definitions/res.Message'
        type: array
      message:
        type: string
      status:
        type: boolean
    type: object
  res.Message:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
...
  schemas.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  schemas.ForgotPasswordVerifyRequest:
    properties:
      confirm_password:
        type: string
      email:
        type: string
      new_password:
        minLength: 8
        type: string
      otp_code:
        type: string
    required:
    - confirm_password
    - email
    - new_password
    - otp_code
    type: object
  schemas.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
      rememberMe:
        type: boolean
    required:
    - email
    - password
    type: object
  schemas.LogoutRequest:
    properties:
      allDevices:
        description: logout from all devices
        type: boolean
    type: object
  schemas.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  schemas.RegisterRequest:
    properties:
      confirmPassword:
        type: string
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
      username:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - confirmPassword
    - email
    - firstName
    - lastName
    - password
    - username
    type: object
...
  schemas.ResendOTPRequest:
    properties:
      email:
        type: string
      otp_type:
        enum:
        - verification
        - forgot_password
        type: string
    required:
    - email
    - otp_type
    type: object
...
  schemas.VerifyOTPRequest:
    properties:
      email:
        type: string
      otp_code:
        type: string
      otp_type:
        enum:
        - verification
        - forgot_password
        type: string
    required:
    - email
    - otp_code
    - otp_type
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: E-Commerce NW SBO Backend API
  title: E-Commerce NW API
  version: "1.0"
paths:
...
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Send OTP for password reset
      parameters:
      - description: Email
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.ErrorConstant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      summary: POST Forgot Password
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login with email and password
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.ErrorConstant'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      summary: POST Login
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logout and revoke session
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Logout options
        in: body
        name: request
        schema:
          $ref: '#/definitions/schemas.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.ErrorConstant'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      summary: POST Logout
      tags:
      - Auth
  /auth/refresh-token:
    post:
      consumes:
      - application/json
      description: Refresh access token using refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.ErrorConstant'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      summary: POST Refresh Token
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user account and send verification OTP
      parameters:
      - description: Registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/res.ErrorConstant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      summary: POST Register
      tags:
      - Auth
  /auth/resend-otp:
    post:
      consumes:
      - application/json
      description: Resend OTP code to email
      parameters:
      - description: Resend OTP data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.ResendOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.ErrorConstant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      summary: POST Resend OTP
      tags:
      - Auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset password using OTP code
      parameters:
      - description: Password reset data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.ForgotPasswordVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.ErrorConstant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      summary: POST Reset Password with OTP
      tags:
      - Auth
  /auth/sessions:
    get:
      consumes:
      - application/json
      description: Get all active sessions for current user
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.ErrorConstant'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      summary: GET Sessions
      tags:
      - Auth
  /auth/sessions/{session_id}:
    delete:
      consumes:
      - application/json
      description: Revoke a specific session
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Session ID
        in: path
        name: session_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.ErrorConstant'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      summary: DELETE Session
      tags:
      - Auth
  /auth/verify-otp:
    post:
      consumes:
      - application/json
      description: Verify OTP code for email verification or password reset
      parameters:
      - description: OTP verification data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.VerifyOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.ErrorConstant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      summary: POST Verify OTP
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"